refactor(connect): move stellar method definitions to corresponding folder
