+++ packages/connect-common/src/types/api/stellar/stellarGetAddress.ts
@@ -1,4 +1,4 @@
-import type { Address, BundledParams, GetAddress, Params, Response } from '../params';
+import type { Address, BundledParams, GetAddress, Params, Response } from '../../params';
 
 export declare function stellarGetAddress(params: Params<GetAddress>): Response<Address>;
 export declare function stellarGetAddress(params: BundledParams<GetAddress>): Response<Address[]>;
