+++ packages/connect-common/src/types/api/stellar/stellarSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { StellarSignTransaction, StellarSignedTx } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function stellarSignTransaction(
+    params: Params<StellarSignTransaction>,
+): Response<StellarSignedTx>;
